Auto-update lookup table
Added a lookup table for the instructions and hex values rather than hardcoded if statements.

When finding the already existing return instructions (before checkinstruction() is called) it now loads the lookup table (not very optimised yet), searches in there and outputs the data if found. If it isn't found, it gathers the information required and appends it to the table file. The tool can then be run again.

Seems to be working okay (based on my knowledge). There are a few things that should probably be changed and cleaned up.